fix: Fret click fix
While clicking a fret that had a mark on it, in a space of a fret and not the mark, fret index was incorrectly calculated leading to invalid data being returned to user defined fretClick callback. Now Mark instance holds callback with correct data which is created through the same "Fret.callback" method with difference of passing "true" as argument to the method which distinguishes Mark from Fret for the sole purpose of proper fret indexing.